refactor done but one test fails
