+++ autolens/point/solver/step.py
@@ -4,12 +4,6 @@ from dataclasses import dataclass
 import autoarray as aa
 from autoarray.numpy_wrapper import register_pytree_node_class
 
-try:
-    from autoarray.structures.triangles.array.jax_array import ArrayTriangles
-except ImportError:
-    from autoarray.structures.triangles.array import ArrayTriangles
-
-
 logger = logging.getLogger(__name__)
