+++ test_autolens/point/triangles/test_solver.py
@@ -1,13 +1,22 @@
-from typing import Tuple
-
 import numpy as np
 import pytest
+import time
+from typing import Tuple
 
-import autolens as al
 import autogalaxy as ag
-from autoarray.structures.triangles.coordinate_array import CoordinateArrayTriangles
+import autofit as af
+from autolens import PointSolver, Tracer
+
+from autoarray.structures.triangles.coordinate_array import (
+    CoordinateArrayTriangles,
+)
+
 from autolens.mock import NullTracer
-from autolens.point.solver import PointSolver
+
+
+@pytest.fixture(autouse=True)
+def register(tracer):
+    af.Model.from_instance(tracer)
 
 
 @pytest.fixture
@@ -18,30 +27,19 @@ def solver(grid):
     )
 
 
-def test_solver_basic(solver):
-    tracer = al.Tracer(
-        galaxies=[
-            al.Galaxy(
-                redshift=0.5,
-                mass=ag.mp.Isothermal(
-                    centre=(0.0, 0.0),
-                    einstein_radius=1.0,
-                ),
-            ),
-            al.Galaxy(
-                redshift=1.0,
-            ),
-        ]
+def test_solver(solver):
+    mass_profile = ag.mp.Isothermal(
+        centre=(0.0, 0.0),
+        einstein_radius=1.0,
     )
-
-    assert solver.solve(
-        tracer=tracer,
+    tracer = Tracer(
+        galaxies=[ag.Galaxy(redshift=0.5, mass=mass_profile)],
+    )
+    result = solver.solve(
+        tracer,
         source_plane_coordinate=(0.0, 0.0),
     )
-
-
-def test_steps(solver):
-    assert solver.n_steps == 7
+    assert result
 
 
 @pytest.mark.parametrize(
@@ -59,32 +57,19 @@ def test_steps(solver):
 def test_trivial(
     source_plane_coordinate: Tuple[float, float],
     grid,
+    solver,
 ):
-    solver = PointSolver.for_grid(
-        grid=grid,
-        pixel_scale_precision=0.01,
-    )
     coordinates = solver.solve(
-        tracer=NullTracer(),
+        NullTracer(),
         source_plane_coordinate=source_plane_coordinate,
     )
-
+    coordinates = coordinates.array[~np.isnan(coordinates.array).any(axis=1)]
     assert coordinates[0] == pytest.approx(source_plane_coordinate, abs=1.0e-1)
 
-
-def triangle_set(triangles):
-    return {
-        tuple(sorted([tuple(np.round(pair, 4)) for pair in triangle]))
-        for triangle in triangles.triangles.tolist()
-        if not np.isnan(triangle).any()
-    }
-
-
-def test_real_example_normal(grid, tracer):
+def test_real_example_jax(grid, tracer):
     jax_solver = PointSolver.for_grid(
         grid=grid,
         pixel_scale_precision=0.001,
-        array_triangles_cls=CoordinateArrayTriangles,
     )
 
     result = jax_solver.solve(
@@ -92,4 +77,4 @@ def test_real_example_normal(grid, tracer):
         source_plane_coordinate=(0.07, 0.07),
     )
 
-    assert len(result) == 5
+    assert len(result) == 15
